refactor: move type.ts
